Revert "mongo connections inside one actor"
