+++ spark-mongodb/src/main/scala/com/stratio/datasource/mongodb/client/MongodbClientFactory.scala
@@ -17,20 +17,20 @@
 package com.stratio.datasource.mongodb.client
 
 import java.util.concurrent._
+import javax.net.ssl.SSLSocketFactory
 
-import akka.actor.{ActorSystem, Props}
-import akka.pattern.ask
-import akka.util.Timeout
+import akka.actor.ActorSystem
 import com.mongodb.ServerAddress
 import com.mongodb.casbah.Imports._
-import com.mongodb.casbah.MongoClient
-import com.stratio.datasource.mongodb.client.MongodbClientActor._
-import com.stratio.datasource.mongodb.config.MongodbSSLOptions
+import com.mongodb.casbah.{MongoClient, MongoClientOptions}
+import com.stratio.datasource.mongodb.config.MongodbConfig.{ReadPreference => ProviderReadPreference, _}
+import com.stratio.datasource.mongodb.config.{MongodbConfig, MongodbSSLOptions}
 import com.typesafe.config.ConfigFactory
 
-import scala.concurrent.Await
+import scala.annotation.tailrec
+import scala.collection.JavaConversions._
 import scala.concurrent.duration._
-
+import scala.util.Try
 
 /**
  * Different client configurations to Mongodb database
@@ -39,34 +39,65 @@ object MongodbClientFactory {
 
   type Client = MongoClient
 
-  private val CloseAttempts = 120
+  val KeySeparator = "-"
+
+  val CloseSleepTime = 1000
+
+  val CloseAttempts = 180
+
+  /**
+   * MongoDb Client connections are saved in a Concurrent hashMap, is used to have only one concurrent connection for
+   * each operation, when this are finished the connection are reused
+   */
+  var mongoClient: scala.collection.concurrent.Map[String, MongodbConnection] =
+    new ConcurrentHashMap[String, MongodbConnection]()
 
   /**
    * Scheduler that close connections automatically when the timeout was expired
    */
-  private val actorSystem = ActorSystem("mongodbClientFactory", ConfigFactory.load(ConfigFactory.parseString("akka.daemonic=on")))
-  private val scheduler = actorSystem.scheduler
-  private val SecondsToCheckConnections = 60
-  private val mongoConnectionsActor = actorSystem.actorOf(Props(new MongodbClientActor), "mongoConnectionActor")
-
-  private implicit val executor = actorSystem.dispatcher
-  private implicit val timeout: Timeout = Timeout(3.seconds)
-
+  val actorSystem = ActorSystem()
+  val scheduler = actorSystem.scheduler
+  val SecondsToCheckConnections = 60
+  val task = new Runnable {
+    def run() {
+      synchronized {
+        val currentTime = System.currentTimeMillis()
+        mongoClient.foreach { case (key, connection) =>
+          if((connection.status == ConnectionStatus.Free) && (connection.timeOut <= currentTime)) {
+            connection.client.close()
+            mongoClient.remove(key)
+          }
+        }
+      }
+    }
+  }
+  implicit val executor = actorSystem.dispatcher
   scheduler.schedule(
     initialDelay = Duration(SecondsToCheckConnections, TimeUnit.SECONDS),
     interval = Duration(SecondsToCheckConnections, TimeUnit.SECONDS),
-    mongoConnectionsActor,
-    CheckConnections)
+    runnable = task)
 
   /**
    * Get or Create one client connection to MongoDb
    * @param host Ip or Dns to connect
-   * @return Client connection with identifier
+   * @return Client connection
    */
-  def getClient(host: String): ClientResponse = {
-    val futureResult = mongoConnectionsActor ? GetClient(host)
-    Await.result(futureResult, timeout.duration) match {
-      case ClientResponse(key, clientConnection) => ClientResponse(key, clientConnection)
+  def getClient(host: String): (String, Client) = {
+    synchronized {
+      val hostPort = new ServerAddress(host)
+      val connKey = connectionKey(0, List(hostPort))
+      val (finalKey, connection) = mongoClient.get(connKey) match {
+        case Some(client) =>
+          if (client.status == ConnectionStatus.Free) (connKey, client)
+          else createClient(connKey, host)
+        case None => createClient(connKey, host)
+      }
+
+      mongoClient.update(finalKey, connection.copy(
+        timeOut = System.currentTimeMillis() + DefaultConnectionsTime,
+        status = ConnectionStatus.Busy))
+
+      (finalKey, connection.client)
     }
   }
 
@@ -77,12 +108,26 @@ object MongodbClientFactory {
    * @param user User for credentials
    * @param database Database for credentials
    * @param password Password for credentials
-   * @return Client connection with identifier
+   * @return Client connection
    */
-  def getClient(host: String, port: Int, user: String, database: String, password: String): ClientResponse = {
-    val futureResult = mongoConnectionsActor ? GetClientWithUser(host, port, user, database, password)
-    Await.result(futureResult, timeout.duration) match {
-      case ClientResponse(key, clientConnection) => ClientResponse(key, clientConnection)
+  def getClient(host: String, port: Int, user: String, database: String, password: String): (String, Client) = {
+    synchronized {
+      val credentials = List(MongoCredential.createCredential(user, database, password.toCharArray))
+      val hostPort = new ServerAddress(host, port)
+      val connKey = connectionKey(0, List(hostPort), credentials)
+
+      val (finalKey, connection) = mongoClient.get(connKey) match {
+        case Some(client) =>
+          if (client.status == ConnectionStatus.Free) (connKey, client)
+          else createClient(1 + connKey, List(hostPort), credentials)
+        case None => createClient(connKey, List(hostPort), credentials)
+      }
+
+      mongoClient.update(finalKey, connection.copy(
+        timeOut = System.currentTimeMillis() + DefaultConnectionsTime,
+        status = ConnectionStatus.Busy))
+
+      (finalKey, connection.client)
     }
   }
 
@@ -92,26 +137,118 @@ object MongodbClientFactory {
    * @param credentials Credentials to connect
    * @param optionSSLOptions SSL options for secure connections
    * @param clientOptions All options for the client connections
-   * @return Client connection with identifier
+   * @return Client connection
    */
   def getClient(hostPort: List[ServerAddress],
                 credentials: List[MongoCredential] = List(),
                 optionSSLOptions: Option[MongodbSSLOptions] = None,
-                clientOptions: Map[String, Any] = Map()): ClientResponse = {
-    val futureResult =
-      mongoConnectionsActor ? GetClientWithMongoDbConfig(hostPort, credentials, optionSSLOptions, clientOptions)
-    Await.result(futureResult, timeout.duration) match {
-      case ClientResponse(key, clientConnection) => ClientResponse(key, clientConnection)
+                clientOptions: Map[String, Any] = Map()): (String, Client) = {
+    synchronized {
+      val connKey = connectionKey(0, hostPort, credentials, clientOptions)
+      val (finalKey, connection) = mongoClient.get(connKey) match {
+        case Some(client) =>
+          if (client.status == ConnectionStatus.Free) (connKey,client)
+          else createClient(1 + connKey, hostPort, credentials, optionSSLOptions, clientOptions)
+        case None => createClient(connKey, hostPort, credentials, optionSSLOptions, clientOptions)
+      }
+
+      mongoClient.update(finalKey, connection.copy(
+        timeOut = System.currentTimeMillis() +
+          extractValue[String](clientOptions, ConnectionsTime).map(_.toLong).getOrElse(DefaultConnectionsTime),
+        status = ConnectionStatus.Busy))
+
+      (finalKey, connection.client)
     }
   }
 
+  private def createClient(key: String , host: String): (String ,MongodbConnection) = {
+    saveConnection(key, MongodbConnection(MongoClient(host)))
+  }
+
+  private def createClient(key: String,
+                           hostPort: List[ServerAddress],
+                           credentials: List[MongoCredential] = List(),
+                           optionSSLOptions: Option[MongodbSSLOptions] = None,
+                           clientOptions: Map[String, Any] = Map()): (String ,MongodbConnection) = {
+
+    val options = {
+
+      val builder = new MongoClientOptions.Builder()
+        .readPreference(extractValue[String](clientOptions, ProviderReadPreference) match {
+          case Some(preference) => parseReadPreference(preference)
+          case None => DefaultReadPreference
+        })
+        .connectTimeout(extractValue[String](clientOptions, ConnectTimeout).map(_.toInt).getOrElse(DefaultConnectTimeout))
+        .connectionsPerHost(extractValue[String](clientOptions, ConnectionsPerHost).map(_.toInt).getOrElse(DefaultConnectionsPerHost))
+        .maxWaitTime(extractValue[String](clientOptions, MaxWaitTime).map(_.toInt).getOrElse(DefaultMaxWaitTime))
+        .threadsAllowedToBlockForConnectionMultiplier(extractValue[String](clientOptions, ThreadsAllowedToBlockForConnectionMultiplier).map(_.toInt).getOrElse(DefaultThreadsAllowedToBlockForConnectionMultiplier))
+
+      if (sslBuilder(optionSSLOptions)) builder.socketFactory(SSLSocketFactory.getDefault())
+
+      builder.build()
+    }
+
+    saveConnection(key, MongodbConnection(MongoClient(hostPort, credentials, options)))
+  }
+
+  @tailrec
+  private def saveConnection(key: String, mongoDbConnection : MongodbConnection) : (String, MongodbConnection) = {
+    mongoClient.putIfAbsent(key, mongoDbConnection) match {
+      case Some(_) =>
+        val splittedKey = key.split(KeySeparator)
+        val index = splittedKey.headOption match {
+          case Some(indexNumber) => Try(indexNumber.toInt + 1).getOrElse(0)
+          case None => 0
+        }
+        saveConnection(s"$index${splittedKey.drop(1).mkString(KeySeparator)}", mongoDbConnection)
+      case None => (key, mongoDbConnection)
+    }
+  }
+
+  /**
+   * Create the connection string for the concurrent hashMap, the params make the unique key
+   * @param index Index for the same concurrent connections to the same database with the same options
+   * @param hostPort List of servers addresses
+   * @param credentials Credentials for connect
+   * @param clientOptions All options for the client connections
+   * @return The calculated string
+   */
+  @tailrec
+  private def connectionKey(index : Int,
+                            hostPort: List[ServerAddress],
+                            credentials: List[MongoCredential] = List(),
+                            clientOptions: Map[String, Any] = Map()): String = {
+    val key = if (clientOptions.nonEmpty)
+      s"$index-${clientOptions.mkString(KeySeparator)}"
+    else s"$index$KeySeparator${hostPort.mkString(KeySeparator)}$KeySeparator${credentials.mkString(KeySeparator)}"
+
+    val clientFound = mongoClient.find { case(clientKey, connection) =>
+      clientKey == key && connection.status == ConnectionStatus.Busy
+    }
+
+    clientFound match {
+      case Some(client) => connectionKey(index + 1, hostPort, credentials, clientOptions)
+      case None => key
+    }
+  }
 
   /**
    * Close all client connections on the concurrent map
    * @param gracefully Close the connections if is free
    */
-  def closeAll(gracefully: Boolean = true, attempts: Int = CloseAttempts): Unit = {
-    mongoConnectionsActor ! CloseAll(gracefully, attempts)
+  def closeAll(gracefully : Boolean = true, attempts : Int = CloseAttempts): Unit = {
+    synchronized {
+      mongoClient.foreach { case (key, connection) =>
+        if (!gracefully || connection.status == ConnectionStatus.Free) {
+          connection.client.close()
+          mongoClient.remove(key)
+        }
+      }
+      if (mongoClient.nonEmpty && attempts > 0) {
+        Thread.sleep(CloseSleepTime)
+        closeAll(gracefully, attempts - 1)
+      }
+    }
   }
 
   /**
@@ -119,8 +256,16 @@ object MongodbClientFactory {
    * @param client client value for connect to MongoDb
    * @param gracefully Close the connection if is free
    */
-  def closeByClient(client: Client, gracefully: Boolean = true): Unit = {
-    mongoConnectionsActor ! CloseByClient(client, gracefully)
+  def close(client: Client, gracefully: Boolean = true): Unit = {
+    synchronized {
+      mongoClient.find { case (key, clientSearch) => clientSearch.client == client }
+        .foreach { case (key, clientFound) =>
+          if (!gracefully || clientFound.status == ConnectionStatus.Free) {
+            clientFound.client.close()
+            mongoClient.remove(key)
+          }
+        }
+    }
   }
 
   /**
@@ -129,29 +274,40 @@ object MongodbClientFactory {
    * @param gracefully Close the connection if is free
    */
   def closeByKey(clientKey: String, gracefully: Boolean = true): Unit = {
-    mongoConnectionsActor ! CloseByKey(clientKey, gracefully)
+    synchronized {
+      mongoClient.get(clientKey).foreach(clientFound => {
+        if (!gracefully || clientFound.status == ConnectionStatus.Free) {
+          clientFound.client.close()
+          mongoClient.remove(clientKey)
+        }
+      })
+    }
   }
 
   /**
    * Set Free the connection that have the same client as the client param
    * @param client client value for connect to MongoDb
    */
-  def setFreeConnectionByClient(client: Client, extendedTime: Option[Long] = None): Unit = {
-    mongoConnectionsActor ! SetFreeConnectionsByClient(client, extendedTime)
+  def setFreeConnection(client: Client, extendedTime : Option[Long] = None): Unit = {
+    synchronized {
+      mongoClient.find { case (key, clientSearch) => clientSearch.client == client }
+        .foreach { case (key, clientFound) =>
+          mongoClient.update(key, clientFound.copy(status = ConnectionStatus.Free,
+            timeOut = System.currentTimeMillis() + extendedTime.getOrElse(DefaultConnectionsTime)))
+        }
+    }
   }
 
   /**
    * Set Free the connection that have the same key as the clientKey param
    * @param clientKey key pre calculated with the connection options
    */
-  def setFreeConnectionByKey(clientKey: String, extendedTime: Option[Long] = None): Unit = {
-    mongoConnectionsActor ! SetFreeConnectionByKey(clientKey, extendedTime)
-  }
-
-  def getSize: Int = {
-    val futureResult = mongoConnectionsActor ? GetSize
-    Await.result(futureResult, timeout.duration) match {
-      case size: Int => size
+  def setFreeConnectionByKey(clientKey: String, extendedTime : Option[Long] = None): Unit = {
+    synchronized {
+      mongoClient.get(clientKey).foreach(clientFound => {
+        mongoClient.update(clientKey, clientFound.copy(status = ConnectionStatus.Free,
+          timeOut = System.currentTimeMillis() + extendedTime.getOrElse(DefaultConnectionsTime)))
+      })
     }
   }
 
@@ -159,7 +315,7 @@ object MongodbClientFactory {
   def extractValue[T](options: Map[String, Any], key: String): Option[T] =
     options.get(key.toLowerCase).map(_.asInstanceOf[T])
 
-  def sslBuilder(optionSSLOptions: Option[MongodbSSLOptions]): Boolean =
+  private def sslBuilder(optionSSLOptions: Option[MongodbSSLOptions]): Boolean =
     optionSSLOptions.exists(sslOptions => {
       if (sslOptions.keyStore.nonEmpty) {
         System.setProperty("javax.net.ssl.keyStore", sslOptions.keyStore.get)
@@ -173,5 +329,4 @@ object MongodbClientFactory {
       }
       true
     })
-
 }
