+++ spark-mongodb/src/main/scala/com/stratio/datasource/mongodb/partitioner/MongodbPartitioner.scala
@@ -61,12 +61,12 @@ class MongodbPartitioner(config: Config) extends Partitioner[MongodbPartition] {
   private val cursorBatchSize = config.getOrElse[Int](MongodbConfig.CursorBatchSize, MongodbConfig.DefaultCursorBatchSize)
 
   override def computePartitions(): Array[MongodbPartition] = {
-    val mongoClient = MongodbClientFactory.getClient(hosts, credentials, ssloptions, clientOptions)
+    val mongoClient = MongodbClientFactory.getClient(hosts, credentials, ssloptions, clientOptions)._2
 
-    val result = if (isShardedCollection(mongoClient.clientConnection))
-      computeShardedChunkPartitions(mongoClient.clientConnection)
+    val result = if (isShardedCollection(mongoClient))
+      computeShardedChunkPartitions(mongoClient)
     else
-      computeNotShardedPartitions(mongoClient.clientConnection)
+      computeNotShardedPartitions(mongoClient)
 
     result
   }
@@ -203,14 +203,14 @@ class MongodbPartitioner(config: Config) extends Partitioner[MongodbPartition] {
           .find(MongoDBObject("_id" -> stats.getString("primary"))).batchSize(cursorBatchSize)
         val shard = shards.next()
         val shardHost: String = shard.as[String]("host").replace(shard.get("_id") + "/", "")
-        val shardClient = MongodbClientFactory.getClient(shardHost)
-        val data = shardClient.clientConnection.getDB("admin").command(cmd)
+        val (shardClientKey, shardClient) = MongodbClientFactory.getClient(shardHost)
+        val data = shardClient.getDB("admin").command(cmd)
         val splitKeys = data.as[List[DBObject]]("splitKeys").map(Option(_))
         val ranges = (splitKeyMin +: splitKeys) zip (splitKeys :+ splitKeyMax )
 
         shards.close()
 
-        MongodbClientFactory.setFreeConnectionByKey(shardClient.key, connectionsTime)
+        MongodbClientFactory.setFreeConnectionByKey(shardClientKey, connectionsTime)
 
         ranges.toSeq
     }.getOrElse(Seq((None, None)))
